improve: navbar design
