+++ src/components/About/About.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import {Section} from "../Resources/WriteUps/WriteUpsElements";
+
+const About = () => {
+    return (
+        <Section>
+            About
+        </Section>
+    );
+};
+
+export default About;
