+++ src/components/Community/Community.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import {Section} from "../Resources/WriteUps/WriteUpsElements";
+
+const Community = () => {
+    return (
+        <Section>
+            Community
+        </Section>
+    );
+};
+
+export default Community;
