+++ src/components/Navbar/Dropdowns/Dropdown.jsx
@@ -2,23 +2,23 @@ import React, {useState} from 'react';
 import styled from "styled-components";
 import SideDropdown from './SideDropdown';
 import { Link } from 'react-router-dom';
-import { NavLinks } from '../Navbar/NavbarElements';
+import {NavLinks, SideDropdownIcon} from '../Navbar/NavbarElements';
 
 export default function Dropdown(props) {
   const menu = [
     {title: "Roadmap", url: "/resources/roadmapResources" },
     {title: "Prep", url: ""},
-    {title: "Events", url:"/resources/events" },
-    {title: "Blogs", url: "/resources/blogs"},
+    {title: "Events", url:"/events" },
+    {title: "Blogs", url: "/blogs"},
     {title: "News", url: "/resources/cyberNews"},
   ]
 
   const [sideDrop, setSideDrop] = useState(false);
 
   const sideDropHandler = (title)=> {
-     props.sidebar ? 
-     title == 'Prep' && setSideDrop(!sideDrop) : 
-     title == 'Prep' ? setSideDrop(true) : setSideDrop(false);
+     props.sidebar ?
+     title === 'Prep' && setSideDrop(!sideDrop) :
+     title === 'Prep' ? setSideDrop(true) : setSideDrop(false);
   }
 
   const Item = styled.ul`
@@ -30,8 +30,18 @@ export default function Dropdown(props) {
   `
 
   const mobStyle = {width :'100%', padding: 6};
-  const dStyle = {width :130, padding: 6, border:'1px solid green', color:'white', position: 'absolute', top: 80 };
-  
+  const dStyle = {
+    borderRadius: 8,
+    background: '#101417',
+    border: '1px solid #101417',
+    width: 130,
+    padding: 6,
+    color: 'white',
+    position: 'absolute',
+    top: 80
+  };
+
+
 
   return (
     <div style={props.sidebar ? mobStyle : dStyle}>
@@ -41,15 +51,15 @@ export default function Dropdown(props) {
                   !props.sidebar ?
                   <Item onMouseEnter={() => sideDropHandler(i.title)}>
                     <Link style={{textDecoration: 'none', color:'white' }} to={i.url}> {i.title} </Link >
-                    {i.title == 'Prep' && sideDrop && <SideDropdown sidebar={false}/>}
-                  </Item> : 
+                    {i.title === 'Prep' && sideDrop && <SideDropdown sidebar={false}/>}
+                  </Item> :
                   <>
                     <Item onClick={() => sideDropHandler(i.title)}>
-                      {i.title != 'Prep' ? 
-                         <Link style={{textDecoration: 'none', color:'white' }} to={i.url}> {i.title} </Link > : 
+                      {i.title !== 'Prep' ?
+                         <Link style={{textDecoration: 'none', color:'white' }} to={i.url}> {i.title} </Link > :
                          <p style={{textDecoration: 'none', color:'white' }}> {i.title} </p > }
                     </Item>
-                    {i.title == 'Prep' && sideDrop && <SideDropdown sidebar={true}/>}
+                    {i.title === 'Prep' && sideDrop && <SideDropdown sidebar={true}/>}
                   </>
                 ))
             }
