+++ src/components/index.js
@@ -10,18 +10,18 @@ export {default as Socials} from './Homepage/Socials/Socials'
 // ResourcesHome //
 export {default as ResourcesHome} from './Resources/ResourcesHome/ResourcesHome'
 export {default as RoadmapResources} from './Resources/RoadmapResources/RoadmapResources'
-export {default as ResourcesNavbar} from './Resources/ResourcesNavbar/ResourcesNavbar'
-export {default as ResourcesSidebar} from './Resources/ResourcesNavbar/ResourcesSidebar/ResourcesSidebar'
 export {default as Courses} from './Resources/Courses/YoutubeCourses'
 export {default as WriteUps} from './Resources/WriteUps/WriteUps'
 export {default as Blogs} from './Resources/Blogs/Blogs'
 export {default as CyberNews} from './Resources/CyberNews/CyberNews'
-export {default as Events} from './Resources/Events/Events'
+export {default as Events} from './Events/Events'
+
+// //
+export {default as Community} from './Community/Community'
+export {default as About} from './About/About'
 
 // Prep //
 export {default as PrepHome} from './Prep/Prep/PrepHome'
-export {default as PrepNavbar} from './Prep/PrepNavbar/PrepNavbar'
-export {default as PrepSidebar} from './Prep/PrepNavbar/PrepSidebar/PrepSidebar'
 export {default as InterviewQuestions} from './Prep/InterviewQuestions/InterviewQuestions'
 export {default as InterviewExperiences} from './Prep/InterviewExperiences/InterviewExperiences'
 export {default as Quiz} from './Prep/Quiz/Quiz'
