+++ src/pages/Resources.jsx
@@ -1,20 +1,13 @@
-import React, {useState} from 'react';
+import React from 'react';
 import {Route, Routes} from "react-router-dom";
 
-import {ResourcesNavbar,ResourcesSidebar,
+import {
     ResourcesHome, RoadmapResources, CyberNews, Blogs, Events
 } from "../components";
 
 const Resources = () => {
-    const [isOpen, setIsOpen] = useState(false);
-    const toggle = () => setIsOpen(!isOpen);
-
     return (
         <div>
-            <navbar>
-                <ResourcesSidebar isOpen={isOpen} toggle={toggle}/>
-                <ResourcesNavbar toggle={toggle}/>
-            </navbar>
             <Routes>
                 <Route exact path={"/"} element={<ResourcesHome/>}/>
                 <Route exact path={"roadmapResources"} element={<RoadmapResources/>}/>
